LayoutTests/imported/w3c/web-platform-tests/css/css-scroll-anchoring/history-restore-anchors-expected.txt


FAIL Verify scroll anchoring interaction with history restoration assert_equals: expected 1100 but got 1000
